adding comments for JavaDoc

+++ src/equipos/Averia.java
@@ -2,7 +2,11 @@ package equipos;
 
 import java.io.Serializable;
 import java.util.Calendar;
-
+/**
+ * Clase que implementa la interfaz serializable, para manejar las posibles averías de los dispositivos
+ * @author Daniel Pulido
+ *
+ */
 public class Averia implements Serializable {
 	
 	/**
@@ -19,6 +23,15 @@ public class Averia implements Serializable {
 	private String descripcion;
 	private boolean resuelta;
 	// ---------------------------------------- CONSTRUCTORES ---------------------------------
+	/**
+	 * Constructor normal, asigna al objeto el siguiente codigo disponible además de los atributos establecidos en los parámetros.
+	 * @param gravedad Dato de tipo entero que establece la gravedad de la avería.
+	 * @param coste Coste de la reparación en un dato de tipo double.
+	 * @param fechaAlta Cadena de tipo String que contiene la fecha de alta de la avería.
+	 * @param fechaRep Cadena de tipo String que contiene la fecha de reparación de la avería.
+	 * @param descripcion Cadena de tipo String que contiene la descripción completa de la avería.
+	 * @param resuelta Boolean que contiene true si la avería ya ha sido resuelta, en caso contrario contiene false.
+	 */
 	public Averia(int gravedad, double coste, String fechaAlta, String fechaRep, String descripcion,
 			boolean resuelta) {
 		this.codigo = Averia.codSig;
@@ -30,6 +43,16 @@ public class Averia implements Serializable {
 		this.descripcion = descripcion;
 		this.resuelta = resuelta;
 	}
+	/**
+	 * Constructor completo, asigna al objeto el  codigo pasado por parámetro además de todos los demás atributos.
+	 * @param codigo Dato de tipo int que establece el codigo único de la avería.
+	 * @param gravedad Dato de tipo entero que establece la gravedad de la avería.
+	 * @param coste Coste de la reparación en un dato de tipo double.
+	 * @param fechaAlta Cadena de tipo String que contiene la fecha de alta de la avería.
+	 * @param fechaRep Cadena de tipo String que contiene la fecha de reparación de la avería.
+	 * @param descripcion Cadena de tipo String que contiene la descripción completa de la avería.
+	 * @param resuelta Boolean que contiene true si la avería ya ha sido resuelta, en caso contrario contiene false.
+	 */
 	public Averia(int codigo, int gravedad, double coste, String fechaAlta, String fechaRep, String descripcion,
 			boolean resuelta) {
 		this.codigo = codigo;
@@ -40,6 +63,10 @@ public class Averia implements Serializable {
 		this.descripcion = descripcion;
 		this.resuelta = resuelta;
 	}
+	/**
+	 * Constructor básico, únicamente asigna el código automático y la gravedad pasada por parámetro, establece como fecha de alta la de ejecución del constructor.
+	 * @param gravedad Entero que establece la gravedad de la avería.
+	 */
 	public Averia(int gravedad) {
 		this.codigo = Averia.codSig;
 		Averia.codSig++;
@@ -50,6 +77,9 @@ public class Averia implements Serializable {
 		this.descripcion = "";
 		this.resuelta = false;
 	}
+	/**
+	 * Constructor por defecto, solo establece el código y la fecha de alta, los demás campos quedan con los valores por defecto.
+	 */
 	public Averia() {
 		this.codigo = Averia.codSig;
 		Averia.codSig++;
@@ -60,6 +90,10 @@ public class Averia implements Serializable {
 		this.descripcion = "";
 		this.resuelta = false;
 	}
+	/**
+	 * Constructor copia, realiza una copia exacta de una avería.
+	 * @param a Objeto de tipo Averia que se copiará en la nueva instancia.
+	 */
 	public Averia(Averia a) {
 		this.codigo = a.codigo;
 		this.gravedad = a.gravedad;
@@ -146,10 +180,18 @@ public class Averia implements Serializable {
 	public void marcarReparada() {
 		this.resuelta = true;
 	}
-	public void subirCoste(int prima) {
+	/**
+	 * aumenta el coste de la reparación, sumandole el importe pasado en el parámetro.
+	 * @param prima double que se sumará al coste.
+	 */
+	public void subirCoste(double prima) {
 		this.coste += prima;
 	}
-	public void bajarCoste(int descuento) {
+	/**
+	 * reduce el coste de la reparación restandole el importe pasado en el parámetro.
+	 * @param descuento Double que se restará del coste.
+	 */
+	public void bajarCoste(double descuento) {
 		this.coste-=descuento;
 	}
 	public void mostrarCoste() {
@@ -162,6 +204,9 @@ public class Averia implements Serializable {
 			System.out.println("La avería "+this.codigo+" está pendiente de reparación");
 		}
 	}
+	/**
+	 * Muestra por consola los datos de la avería
+	 */
 	public void mostrarAveria() {
 		System.out.println("Código averia: "+ this.codigo);
 		System.out.println("Descripción: "+this.descripcion);
